dataset/requirements.txt
earthengine-api>=0.1.380
requests>=2.31
pandas>=2.0
numpy>=1.24
Pillow>=10.0
